implement uniform triangle sampling
needs to be modified!!!
(changes needed depend on patch subdivision)

currently the full mesh triangles are sampled, not the individual
patches (which, of course, is wrong). we can change this as soon as we
have finished the generation of patches.

lines of code that changed: 470-542
the rest is due to auto-formatting the document.